Fix typos and add better welcome message for no repos
Get the actual setting for which root path is used and display that in
the UI for when there's no repos to be found.

Add the reactive-var package as that makes things easy to get reactive.
Since the return value of the Meteor method is only async on the client
it cannot just be called from the helper.