Remove xml dependency
We now ask hg to give us the default JSON template of the log output. This means we can simplify our parsing and remove the dependency to a xml parser. This has some side effects. The default JSON template does not contain information on how a file was changed in the commit (added/remove etc) so that will not be visible in the log view for now.

Also moving hg_log.js back to the server directory.