Limit results sent to client
Add a query parameter, `maxResults` that will be sent down to the server
on subscription. That will, in turn, limit the number of results
returned from the rx query.

However, right now this is only limiting the first batch of results. If
more results come in later, reactively, those will be added on top of
the limited results.